fix bug

+++ scrapy_redis/extensions/duperfilter.py
@@ -10,28 +10,34 @@ class RFPDupeFilter(BaseDupeFilter):
     """Request Fingerprint duplicates filter"""
     def __init__(self, settings):
         self.logger = logging.getLogger('Dupefilter')
-        redis_conf = settings.get("REDIS_CONF")
-        self.urls_seen = pyreBloom.pyreBloom(
-            "bfilter:{}".format(settings.get('TASK_ID')),
-            100000000, 0.001,
-            host=redis_conf['host'],
-            port=redis_conf['port'],
-            db=redis_conf['db']
-        )
-
+        self.settings = settings
+        self.urls_seen = None
         self.logdupes = True
 
     @classmethod
     def from_settings(cls, settings):
         return cls(settings)
 
+    def open(self, spider):
+        redis_conf = self.settings.get("REDIS_CONF")
+        self.urls_seen = pyreBloom.pyreBloom(
+            "bfilter:{}".format(spider.name),
+            100000000, 0.001,
+            host=redis_conf['host'],
+            port=redis_conf['port'],
+            db=redis_conf['db']
+        )
+
     def request_seen(self, request):
         _uid = md5(request.url).hexdigest()
         if _uid in self.urls_seen:
             return True
 
+        self.urls_seen.add(_uid)
+
     def close(self, reason):
-        self.urls_seen.delete()
+        if self.urls_seen:
+            self.urls_seen.delete()
 
     def log(self, request, spider):
         if self.logdupes:
